app/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.19.2
requests==2.27.1
SQLAlchemy==1.4.45
urllib3==1.26.8
Werkzeug==2.2.2
WTForms==3.0.1